autofac and icon improvements
started implementing IOC (autofac). Only done the window manager so far
but it all loads and stuff so just need to start progressively IOC-ing
all the other classes.
Have also improved the shell icon stuff so that they are much higher
detail. I used someone else's example here but modified it a lot. I need
to add in an attribution to the original coder. That's coming